+++ lib/src/parsing/parser/parser.dart
@@ -11,6 +11,7 @@ import '../../extensions/remove.dart';
 import '../../extensions/show_control_codes.dart';
 import '../../extensions/show_escape_codes.dart';
 import '../../internal/sgr_rules.dart';
+import '../../internal/strings.dart';
 import '../../ready_to_use/csi.dart';
 import '../../ready_to_use/esc.dart';
 import '../../ready_to_use/osc.dart';